@@ -6,18 +6,20 @@ authors = [
     {name = "Your Name", email = "your.email@example.com"}
 ]
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
-    "fastapi==0.68.0",
-    "uvicorn==0.15.0",
-    "alembic==1.6.5",
-    "sqlalchemy==1.3.24",
-    "databases==0.4.3",
-    "asyncpg==0.24.0",
-    "psycopg2==2.9.1",
-    "redis==4.5.4",
-    "pydantic[email]==1.8.2",
-    "orjson==3.6.1"
+    "fastapi>=0.104.0",
+    "uvicorn>=0.24.0",
+    "alembic>=1.12.0",
+    "sqlalchemy>=2.0.0",
+    "databases>=0.8.0",
+    "asyncpg>=0.28.0",
+    "psycopg2-binary>=2.9.0",
+    "redis>=5.0.0",
+    "pydantic>=2.5.0",
+    "orjson>=3.9.0",
+    "pydantic-settings (>=2.10.1,<3.0.0)",
+    "email-validator>=2.0.0"
 ]
 
 [project.optional-dependencies]
